+++ app/build.gradle
@@ -82,5 +82,8 @@ dependencies {
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.14.2'
     kapt 'com.github.bumptech.glide:compiler:4.14.2'
+
+    //pulltorefresh
+    implementation "androidx.swiperefreshlayout:swiperefreshlayout:1.1.0"
 }
