+++ app/src/main/java/com/example/instachat/feature/home/HomeViewModel.kt
@@ -2,21 +2,40 @@ package com.example.instachat.feature.home
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.instachat.services.firebase.FirebaseDataInjector
 import com.example.instachat.services.models.PostModel
 import com.example.instachat.services.models.PostModelItem
+import com.example.instachat.services.repository.RestApiRepository
 import com.google.firebase.firestore.DocumentChange
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.google.gson.Gson
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 @HiltViewModel
-class HomeViewModel @Inject constructor(): ViewModel() {
+class HomeViewModel @Inject constructor(val restApiRepository: RestApiRepository): ViewModel() {
 
     val adapter =  HomeDataAdapter()
+
+    fun injectPostsToFirebase() {
+        viewModelScope.launch (Dispatchers.IO){
+            val imageListResponse = restApiRepository.lorealImageListRestClient.getAllImages()
+            val postsList = restApiRepository.jsonPlaceHolderApiClient.getAllPosts()
+
+            postsList.forEach { post ->
+                post.postImageUrl = imageListResponse.get(post.id - 1).download_url
+                viewModelScope.launch(Dispatchers.IO) {
+                    FirebaseDataInjector.injectPostsToFirebase(post)
+                }
+            }
+        }
+        getAllPosts()
+    }
 
     fun getAllPosts(){
         val db = Firebase.firestore
