+++ app/src/main/res/layout/activity_main.xml
@@ -8,6 +8,8 @@
         android:layout_height="match_parent"
         tools:context=".MainActivity">
 
+
+
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/container"
             android:name="androidx.navigation.fragment.NavHostFragment"
